+++ internal/cli/clu2adv/clu2adv.go
@@ -10,8 +10,8 @@ func Builder() *cobra.Command {
 	o := &opts{fs: afero.NewOsFs()}
 	cmd := &cobra.Command{
 		Use:     "clusterToAdvancedCluster",
-		Short:   "Convert cluster to advanced_cluster v2",
-		Long:    "Convert a Terraform configuration from mongodbatlas_cluster to mongodbatlas_advanced_cluster schema v2",
+		Short:   "Convert cluster to advanced_cluster preview provider 2.0.0",
+		Long:    "Convert a Terraform configuration from mongodbatlas_cluster to mongodbatlas_advanced_cluster preview provider 2.0.0",
 		Aliases: []string{"clu2adv"},
 		RunE: func(_ *cobra.Command, _ []string) error {
 			if err := o.PreRun(); err != nil {