+++ internal/convert/const_names.go
@@ -51,4 +51,7 @@ const (
 	nMoved                      = "moved"
 	nFrom                       = "from"
 	nTo                         = "to"
+	nDynamic                    = "dynamic"
+	nForEach                    = "for_each"
+	nContent                    = "content"
 )