+++ internal/hcl/hcl.go
@@ -3,6 +3,7 @@ package hcl
 import (
 	"fmt"
 	"strconv"
+	"strings"
 
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclsyntax"
@@ -34,6 +35,11 @@ func PopAttr(body *hclwrite.Body, attrName, errPrefix string) (hclwrite.Tokens, 
 func SetAttrExpr(body *hclwrite.Body, attrName, expresion string) {
 	tokens := hclwrite.Tokens{{Type: hclsyntax.TokenIdent, Bytes: []byte(expresion)}}
 	body.SetAttributeRaw(attrName, tokens)
+}
+
+// GetAttrExpr returns the expression of an attribute as a string.
+func GetAttrExpr(attr *hclwrite.Attribute) string {
+	return strings.TrimSpace(string(attr.Expr().BuildTokens(nil).Bytes()))
 }
 
 // SetAttrInt sets an attribute to a number.
@@ -79,22 +85,14 @@ func GetAttrString(attr *hclwrite.Attribute, errPrefix string) (string, error) {
 
 // TokensArray creates an array of objects.
 func TokensArray(bodies []*hclwrite.Body) hclwrite.Tokens {
-	ret := hclwrite.Tokens{
-		{Type: hclsyntax.TokenOBrack, Bytes: []byte("[")},
-		{Type: hclsyntax.TokenNewline, Bytes: []byte("\n")},
-	}
+	ret := hclwrite.Tokens{tokenNewLine}
+	tokens := make([]hclwrite.Tokens, 0)
 	for i := range bodies {
-		ret = append(ret, TokensObject(bodies[i])...)
-		if i < len(bodies)-1 {
-			ret = append(ret,
-				&hclwrite.Token{Type: hclsyntax.TokenComma, Bytes: []byte(",")},
-				&hclwrite.Token{Type: hclsyntax.TokenNewline, Bytes: []byte("\n")})
-		}
+		tokens = append(tokens, TokensObject(bodies[i]))
 	}
-	ret = append(ret,
-		&hclwrite.Token{Type: hclsyntax.TokenNewline, Bytes: []byte("\n")},
-		&hclwrite.Token{Type: hclsyntax.TokenCBrack, Bytes: []byte("]")})
-	return ret
+	ret = append(ret, joinTokens(tokens...)...)
+	ret = append(ret, tokenNewLine)
+	return encloseBrackets(ret)
 }
 
 // TokensArraySingle creates an array of one object.
@@ -104,15 +102,31 @@ func TokensArraySingle(body *hclwrite.Body) hclwrite.Tokens {
 
 // TokensObject creates an object.
 func TokensObject(body *hclwrite.Body) hclwrite.Tokens {
-	tokens := RemoveLeadingNewline(body.BuildTokens(nil))
-	ret := hclwrite.Tokens{
-		{Type: hclsyntax.TokenOBrace, Bytes: []byte("{")},
-		{Type: hclsyntax.TokenNewline, Bytes: []byte("\n")},
-	}
-	ret = append(ret, tokens...)
-	ret = append(ret,
-		&hclwrite.Token{Type: hclsyntax.TokenCBrace, Bytes: []byte("}")})
-	return ret
+	tokens := hclwrite.Tokens{tokenNewLine}
+	tokens = append(tokens, RemoveLeadingNewline(body.BuildTokens(nil))...)
+	return encloseBraces(tokens)
+}
+
+// TokensFromExpr creates the tokens for an expression provided as a string.
+func TokensFromExpr(expr string) hclwrite.Tokens {
+	return hclwrite.Tokens{{Type: hclsyntax.TokenIdent, Bytes: []byte(expr)}}
+}
+
+// TokensObjectFromExpr creates an object with an expression.
+func TokensObjectFromExpr(expr string) hclwrite.Tokens {
+	tokens := hclwrite.Tokens{tokenNewLine}
+	tokens = append(tokens, TokensFromExpr(expr)...)
+	tokens = append(tokens, tokenNewLine)
+	return encloseBraces(tokens)
+}
+
+// TokensFuncMerge creates the tokens for the HCL merge function.
+func TokensFuncMerge(tokens ...hclwrite.Tokens) hclwrite.Tokens {
+	params := hclwrite.Tokens{tokenNewLine}
+	params = append(params, joinTokens(tokens...)...)
+	params = append(params, tokenNewLine)
+	ret := hclwrite.Tokens{{Type: hclsyntax.TokenIdent, Bytes: []byte("merge")}}
+	return append(ret, encloseParens(params)...)
 }
 
 // RemoveLeadingNewline removes the first newline if it exists to make the output prettier.
@@ -133,9 +147,46 @@ func AppendComment(body *hclwrite.Body, comment string) {
 
 // GetParser returns a parser for the given config and checks HCL syntax is valid
 func GetParser(config []byte) (*hclwrite.File, error) {
-	parser, diags := hclwrite.ParseConfig(config, "", hcl.Pos{Line: 1, Column: 1})
+	parser, diags := hclwrite.ParseConfig(config, "", hcl.InitialPos)
 	if diags.HasErrors() {
 		return nil, fmt.Errorf("failed to parse Terraform config file: %s", diags.Error())
 	}
 	return parser, nil
+}
+
+var tokenNewLine = &hclwrite.Token{Type: hclsyntax.TokenNewline, Bytes: []byte("\n")}
+
+// joinTokens joins multiple tokens with commas and newlines.
+func joinTokens(tokens ...hclwrite.Tokens) hclwrite.Tokens {
+	ret := hclwrite.Tokens{}
+	for i := range tokens {
+		ret = append(ret, tokens[i]...)
+		if i < len(tokens)-1 {
+			ret = append(ret,
+				&hclwrite.Token{Type: hclsyntax.TokenComma, Bytes: []byte(",")},
+				tokenNewLine)
+		}
+	}
+	return ret
+}
+
+// encloseParens encloses tokens with parentheses, ( ).
+func encloseParens(tokens hclwrite.Tokens) hclwrite.Tokens {
+	ret := hclwrite.Tokens{{Type: hclsyntax.TokenOParen, Bytes: []byte("(")}}
+	ret = append(ret, tokens...)
+	return append(ret, &hclwrite.Token{Type: hclsyntax.TokenCParen, Bytes: []byte(")")})
+}
+
+// encloseBraces encloses tokens with curly braces, { }.
+func encloseBraces(tokens hclwrite.Tokens) hclwrite.Tokens {
+	ret := hclwrite.Tokens{{Type: hclsyntax.TokenOBrace, Bytes: []byte("{")}}
+	ret = append(ret, tokens...)
+	return append(ret, &hclwrite.Token{Type: hclsyntax.TokenCBrace, Bytes: []byte("}")})
+}
+
+// encloseBrackets encloses tokens with square brackets, [ ].
+func encloseBrackets(tokens hclwrite.Tokens) hclwrite.Tokens {
+	ret := hclwrite.Tokens{{Type: hclsyntax.TokenOBrack, Bytes: []byte("[")}}
+	ret = append(ret, tokens...)
+	return append(ret, &hclwrite.Token{Type: hclsyntax.TokenCBrack, Bytes: []byte("]")})
 }