+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## (Unreleased)
 
+ENHANCEMENTS:
+
+* Supports `dynamic` block for `tags` and `labels`
+
 ## 1.0.0 (Mar 6, 2025)
 
 NOTES: